theme: add layout files for themes directory
this way we can move themes into a subdirectory themes without the need to define 'docs' as layout. the single pages have a table of contents on the left and all documentation to one single theme is in one single page bundle. better for seo have a look at gethugothemes/documentation for details on page structure. going to move a sample theme over right now, the other themes then successive.